Manny: rename pages directory
