+++ js/src/personal/manny/Manny.page.tsx
@@ -1,7 +1,7 @@
 import { MantineProvider } from '@mantine/core'
 // import styles from './Manny.module.css'
 import { theme } from './theme.ts'
-import { Hero } from '@/personal/manny/pages/Hero.tsx'
+import { Hero } from '@/personal/manny/sections/Hero.tsx'
 import { HeaderSimple } from '@/personal/manny/components/HeaderSimple.tsx'
 
 export function MannyPage() {
